requirements.txt
pytesseract
Pillow
pygetwindow
pyautogui
PyQt5
openai
requests
google-generativeai
opencv-python 